+++ MeteorDefense/src/com/esnefedroetem/meteordefense/model/Projectile.java
@@ -6,18 +6,18 @@ import com.esnefedroetem.meteordefense.util.*;
 
 public class Projectile extends MoveableGameObject{
 
-	public Projectile(double angle){
+	public Projectile(float angle){
 		super(angle, Constants.DEFAULT_PROJECTILE_DAMAGE, Constants.DEFAULT_PROJECTILE_SIZE, Constants.DEFAULT_PROJECTILE_SPEED, Constants.DEFAULT_PROJECTILE_SPAWN);
 	}
 	
-	public Projectile(double angle, int damage){
+	public Projectile(float angle, int damage){
 		super(angle, damage, Constants.DEFAULT_PROJECTILE_SIZE, Constants.DEFAULT_PROJECTILE_SPEED, Constants.DEFAULT_PROJECTILE_SPAWN);
 	}
 	
-	public Projectile(double angle, int damage, float size){
+	public Projectile(float angle, int damage, float size){
 		super(angle, damage, Constants.DEFAULT_PROJECTILE_SIZE, Constants.DEFAULT_PROJECTILE_SPEED, Constants.DEFAULT_PROJECTILE_SPAWN);
 	}
-	public Projectile(double angle, int damage, float size, float speed){
+	public Projectile(float angle, int damage, float size, float speed){
 		super(angle, damage, size, speed, Constants.DEFAULT_PROJECTILE_SPAWN);
 	}
